[Feat] auth와 role 타입 분기
