+++ src/types/user.type.ts
@@ -1,0 +1,9 @@
+export interface User {
+  id: string;
+  email: string;
+  name: string;
+  nickname: string;
+  profile_url: string;
+  auth_provider: string;
+  role_id: number;
+}
